+++ elide-graphql/src/test/java/com/yahoo/elide/graphql/PersistentResourceFetcherTest.java
@@ -31,8 +31,10 @@ import java.io.InputStreamReader;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
+import java.util.Date;
 import java.util.HashSet;
 import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 
 /**
@@ -89,12 +91,14 @@ public abstract class PersistentResourceFetcherTest extends GraphQLTest {
         book1.setTitle("Libro Uno");
         book1.setAuthors(new ArrayList<>(Collections.singletonList(author1)));
         book1.setPublisher(publisher1);
+        book1.setPublicationDate(new Date(1514397817135L));
 
         Book book2 = new Book();
         book2.setId(2L);
         book2.setTitle("Libro Dos");
         book2.setAuthors(new ArrayList<>(Collections.singletonList(author1)));
         book2.setPublisher(publisher1);
+        book2.setPublicationDate(new Date(0L));
 
         author1.setPenName(authorOne);
         author1.setBooks(new ArrayList<>(Arrays.asList(book1, book2)));
@@ -128,9 +132,15 @@ public abstract class PersistentResourceFetcherTest extends GraphQLTest {
     }
 
     protected void assertQueryEquals(String graphQLRequest, String expectedResponse) throws Exception {
+        assertQueryEquals(graphQLRequest, expectedResponse, Collections.EMPTY_MAP);
+    }
+
+    protected void assertQueryEquals(String graphQLRequest, String expectedResponse, Map<String, Object> variables) throws Exception {
         boolean isMutation = graphQLRequest.startsWith("mutation");
 
-        ExecutionResult result = api.execute(graphQLRequest, requestScope);
+        ExecutionResult result = api.execute(graphQLRequest, requestScope, variables);
+        // NOTE: We're forcing commit even in case of failures. GraphQLEndpoint tests should ensure we do not commit on
+        //       failure.
         if (isMutation) {
             requestScope.saveOrCreateObjects();
         }
@@ -192,6 +202,12 @@ public abstract class PersistentResourceFetcherTest extends GraphQLTest {
 
     public void runComparisonTest(String testName) throws Exception {
         runComparisonTest(testName, this::assertQueryEquals);
+    }
+
+    public void runComparisonTestWithVariables(String testName, Map<String, Object> variables) throws Exception {
+        String graphQLRequest = loadGraphQLRequest(testName + ".graphql");
+        String graphQLResponse = loadGraphQLResponse(testName + ".json");
+        assertQueryEquals(graphQLRequest, graphQLResponse, variables);
     }
 
     public void runErrorComparisonTest(String testName, String expectedMessage) throws Exception {